+++ src/console/index.ts
@@ -1,0 +1,4 @@
+export {
+  consoleImage,
+  getBase64FromUrl,
+} from './image';